refactor: 유니크 ID 반환 함수 다른 파일로 분리
