+++ src/utils/restaurant.js
@@ -1,0 +1,3 @@
+export function generateUniqueId() {
+  return crypto.randomUUID();
+}
